Upgrade to SDL2

+++ driver/ios/ViewController.m
@@ -149,7 +149,7 @@
 - (void)update
 {
     // XXX: Not sure where this should go.
-    float aspect = fabsf(self.view.bounds.size.width / self.view.bounds.size.height);
+    float aspect = fabs(self.view.bounds.size.width / self.view.bounds.size.height);
     GLKMatrix4 projectionMatrix = GLKMatrix4MakePerspective(GLKMathDegreesToRadians(65.0f), aspect, 0.1f, 100.0f);
     
     // Compute the model view matrix for the object rendered with ES2
